Don't call print stack trace
It was not great that I was printing errors in a number of places. The
fact that I haven't heard anything about this or ever seen these errors
shows that they are in fact very uncommon, however, we need to be able
to disable them at runtime if they annoy users.