Use FindFirstFileExW in native listener
An sbt user discovered that files containing special characters were not
handled correctly by the windows native directory lister. It turns out
that this was because we were not using the wide character apis. Since
the special characters were outside of the ascii range, they'd get
truncated to some value less 32 which is rejected as an invalid path
character by java in Paths.get. Switching to the wide character apis
fixes this issue.