Fix directory watching for missing paths
It was discovered as part of a PR to switch metals to using swoval that
watching a directory that didn't yet exist didn't work on linux and
windows. This was because the watch service was not registering a watch
with newly discovered discoveries. We didn't have this problem with osx
because the fsevents api doesn't care if the directory exists when a
watch is created.