Clarify missing file test case
In #146, it was reported that missing files were not correctly tracked
by the file cache. This was a bit of a misunderstanding because when a
file is registered with the file cache, its parent is not automatically
registered. The parent is monitored for watch events if the file doesn't
exist so that we can detect if the file is created but listing the
parent directory will not work. The file has to be directly queried.