Update register results and log warnings
On osx, there can be situations in which the OS fails to register a
watch stream which can break functionality. When this happens, we want
to return an error to the caller and log a warning since this is very
likely to break the application. There is an argument that we should
actually bomb out on an error rather than just log a warning, but that
seems a bit aggressive.