Fix spurious callbacks from running on registration
Issue #155 reported that there are spurious callbacks run by
the path watcher when registering a prefix of a path when the
suffix had previously been registered. There are internal
callbacks that need to run within the NioPathWatcher to update
its file watch state but we should probably not be running the
creation event callbacks. A simpler fix may have been to just
delete the `runCallbacks(events)` line but that line has
existed pretty much since the begining of the library so I took
the slightly more conservative approach of just filtering out
creation events. That seems to resolve the specific issue
reported in #155.