Add support for filtering paths in FileTreeRepository
There is a bit of an oversight in the existing library that the
FileTreeRepository does not currently permit users to filter for
specific files. This can cause excessive memory utilization tracking
state for files that the user doesn't care about. There also is a lot of
extra work being done behind the scenes to apply the converter function
(which might, for example, hash the file contents) for unneeded files.
The underlying data structures backing the FileTreeRepository support
filtering so I'm not sure why I left that functionality out in the past.

This came up in the context of #155 and
https://github.com/scalameta/metals/pull/3758#pullrequestreview-922146225