Fix overupdate bug
It was reported that the performance of swoval degrade in 1.2.8 due to a
change that caused swoval to fully rescan each directory that was
registered every time. This commit readds the guard on calling update
that was previously present only calling it on the deepest found path in the
directory tree.