Attempt to limit overupdating a second time
The previous attempt at improving the performance for windows failed. It
appears that the pathwatcher may have still been spuriously rescanning
directories in certain situations. This commit updates the logic to
short circuit so that if we detect that the existing watches already
cover a newly registered path that we don't blindly rescan the whole
directory true.